Moved classes to their own file
