+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "cactpot-solver",
-  "version": "1.1.1",
+  "version": "1.1.2",
   "description": "A FFXIV Mini cactpot solver",
   "keywords": [
     "ffxiv",
